refactor(ISV-5857): move merging to sbom.merge
