+++ tests/sbom/test_merge_data/cyclonedx/cachi2.bom.json
