+++ tests/sbom/test_merge_data/cyclonedx/syft-sboms/gomod-pandemonium.bom.json
