+++ tests/sbom/test_merge_data/cyclonedx/syft-sboms/npm-cachi2-smoketest.bom.json
