+++ tests/sbom/test_merge_data/cyclonedx/syft-sboms/pip-e2e-test.bom.json
