+++ tests/sbom/test_merge_data/cyclonedx/syft-sboms/ubi-micro.bom.json
