+++ tests/sbom/test_merge_data/cyclonedx/syft.merged-by-syft.bom.json
