+++ tests/sbom/test_merge_data/cyclonedx/syft.merged-by-us.bom.json
