+++ tests/sbom/test_merge_data/spdx/syft-sboms/gomod-pandemonium.bom.json
