+++ tests/sbom/test_merge_data/spdx/syft-sboms/npm-cachi2-smoketest.bom.json
