+++ tests/sbom/test_merge_data/spdx/syft-sboms/pip-e2e-test.bom.json
