+++ tests/sbom/test_merge_data/spdx/syft-sboms/ubi-micro.bom.json
