+++ tests/sbom/test_merge_data/spdx/syft.merged-by-syft.bom.json
